Fix MenuAppBar
