--- canteenFrontend/src/App.css
@@ -1,19 +1,0 @@
-.navbar {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  margin-top: -170px;
-  margin-left: -300px;
-  background-color: aqua;
-  width: 100%;
-}
-
-.App {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  margin-top: 150px;
-  margin-left: -300px;
-  background-color: blanchedalmond;
-  width: 100%;
-}
