+++ canteenFrontend/src/App.js
@@ -2,22 +2,24 @@ import React from "react";
 import MenuAppBar from "./components/MenuAppBar";
 import ButtonAppBar from "./components/ButtonAppBar";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
-import HomePanel from "./layouts/HomePanel";
-import OrdersPendingPanel from "./layouts/OrdersPendingPanel";
-import OrdersCompletedPanel from "./layouts/OrdersCompletedPanel";
+import HomePanel from "./pages/HomePanel";
+import OrdersPendingPanel from "./pages/OrdersPendingPanel";
+import OrdersCompletedPanel from "./pages/OrdersCompletedPanel";
 
 function App() {
   return (
-    <div className="app">
+    <React.Fragment>
       <Router>
         <div>
-          <Route exact path="/" component={HomePanel} />
-          <Route path="/Orders/pending" component={OrdersPendingPanel} />
-          <Route path="/Orders/completed" component={OrdersCompletedPanel} />
-          <Route path="/Home" component={HomePanel} />
+            <MenuAppBar />
+
+            <Route exact path="/" component={HomePanel} />
+            <Route path="/Orders/pending" component={OrdersPendingPanel} />
+            <Route path="/Orders/completed" component={OrdersCompletedPanel} />
+            <Route path="/Home" component={HomePanel} />
         </div>
       </Router>
-    </div>
+    </React.Fragment>
   );
 }
