+++ canteenFrontend/src/components/MenuAppBar.js
@@ -1,32 +1,28 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core/styles";
+
 import AppBar from "@material-ui/core/AppBar";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
 import IconButton from "@material-ui/core/IconButton";
-import MenuIcon from "@material-ui/icons/Menu";
+import AppIcon from "@material-ui/icons/Fastfood";
 import AccountCircle from "@material-ui/icons/AccountCircle";
 import MenuItem from "@material-ui/core/MenuItem";
 import Menu from "@material-ui/core/Menu";
 
 const styles = {
-  root: {
-    flexGrow: 1
-  },
   grow: {
     flexGrow: 1
   },
+
   appBar: {
-    // width: `100%`,
-    position: "fixed",
-    height: 70,
     backgroundColor: "#0477BD"
   },
-  menuButton: {
-    marginLeft: -12,
-    marginRight: 20
-  }
+
+  toolbar: {
+    height: "72px",
+  },
 };
 
 class MenuAppBar extends React.Component {
@@ -52,56 +48,54 @@ class MenuAppBar extends React.Component {
     const open = Boolean(anchorEl);
 
     return (
-        <div className={classes.root}>
-          <AppBar position="static" className={classes.appBar}>
-            <Toolbar>
-              <IconButton
-                className={classes.menuButton}
-                color="inherit"
-                aria-label="Menu"
-              >
-                <MenuIcon />
-                {/* logo */}
-              </IconButton>
-              <Typography
-                variant="h6"
-                align="right"
-                color="inherit"
-                className={classes.grow}
-              >
-                john
-              </Typography>
+      <AppBar position="sticky">
+        <Toolbar className={classes.toolbar}>
+          <IconButton
+            className={classes.menuButton}
+            color="inherit"
+            aria-label="Menu"
+          >
+            <AppIcon />
+          </IconButton>
 
-              <div>
-                <IconButton
-                  aria-owns={open ? "menu-appbar" : undefined}
-                  aria-haspopup="true"
-                  onClick={this.handleMenu}
-                  color="inherit"
-                >
-                  <AccountCircle fontSize="large" />
-                </IconButton>
+          <Typography
+            variant="h6"
+            align="right"
+            color="inherit"
+            className={classes.grow}
+          >
+            john
+          </Typography>
 
-                <Menu
-                  id="menu-appbar"
-                  anchorEl={anchorEl}
-                  anchorOrigin={{
-                    vertical: "top",
-                    horizontal: "right"
-                  }}
-                  transformOrigin={{
-                    vertical: "top",
-                    horizontal: "right"
-                  }}
-                  open={open}
-                  onClose={this.handleClose}
-                >
-                  <MenuItem onClick={this.handleClose}>Logout</MenuItem>
-                </Menu>
-              </div>
-            </Toolbar>
-          </AppBar>
-        </div>
+          <div>
+            <IconButton
+              aria-owns={open ? "menu-appbar" : undefined}
+              aria-haspopup="true"
+              onClick={this.handleMenu}
+              color="inherit"
+            >
+              <AccountCircle fontSize="large" />
+            </IconButton>
+
+            <Menu
+              id="menu-appbar"
+              anchorEl={anchorEl}
+              anchorOrigin={{
+                vertical: "top",
+                horizontal: "right"
+              }}
+              transformOrigin={{
+                vertical: "top",
+                horizontal: "right"
+              }}
+              open={open}
+              onClose={this.handleClose}
+            >
+              <MenuItem onClick={this.handleClose}>Logout</MenuItem>
+            </Menu>
+          </div>
+        </Toolbar>
+      </AppBar>
     );
   }
 }
