+++ canteenFrontend/src/pages/HomePanel.js
@@ -2,18 +2,14 @@ import React from "react";
 import MenuAppBar from "../components/MenuAppBar";
 import ButtonAppBar from "../components/ButtonAppBar";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
-import OrdersPendingPanel from "../layouts/OrdersPendingPanel";
-import OrdersCompletedPanel from "../layouts/OrdersCompletedPanel";
+import OrdersPendingPanel from "./OrdersPendingPanel";
+import OrdersCompletedPanel from './OrdersCompletedPanel';
 import PermanentDrawerLeft from "../components/PermanentDrawerLeft";
 
 function Home() {
   return (
     <React.Fragment>
-      <div style={{ width: "100%", height: 140 }}>
-        <MenuAppBar />
-        <br />
-        <ButtonAppBar />
-      </div>
+      <ButtonAppBar />
       <PermanentDrawerLeft />
 
       <Route path="/Orders/pending" component={OrdersPendingPanel} />
